convert script
