+++ module/skill/GetUp/src/GetUp.hpp
@@ -2,9 +2,10 @@
 #define MODULE_SKILL_GETUP_HPP
 
 #include <nuclear>
+#include <string>
+#include <vector>
 
 #include "extension/Behaviour.hpp"
-#include "extension/behaviour/Script.hpp"
 
 namespace module::skill {
 
@@ -13,17 +14,17 @@ namespace module::skill {
         /// @brief Stores configuration values
         struct Config {
             /// @brief Script sequence to run when getting from lying on the front to standing
-            std::vector<::extension::behaviour::ScriptRequest> getup_front;
+            std::vector<std::string> getup_front;
             /// @brief Script sequence to run when getting from lying on the back to standing
-            std::vector<::extension::behaviour::ScriptRequest> getup_back;
+            std::vector<std::string> getup_back;
             /// @brief Script sequence to run when getting from lying on the right to standing
-            std::vector<::extension::behaviour::ScriptRequest> getup_right;
+            std::vector<std::string> getup_right;
             /// @brief Script sequence to run when getting from lying on the left to standing
-            std::vector<::extension::behaviour::ScriptRequest> getup_left;
+            std::vector<std::string> getup_left;
             /// @brief Script sequence to run when already upright
-            std::vector<::extension::behaviour::ScriptRequest> getup_upright;
+            std::vector<std::string> getup_upright;
             /// @brief Script sequence to run when told to get up while upside down
-            std::vector<::extension::behaviour::ScriptRequest> getup_upside_down;
+            std::vector<std::string> getup_upside_down;
         } cfg;
 
     public:
